🔨 #34 유저데이터, 플레이리스트 데이터 가져오는 로직 훅으로 분리
